+++ src/main/java/com/postech30/hackathon/service/impl/ClientServiceImpl.java
@@ -59,6 +59,10 @@ public class ClientServiceImpl implements ClientService {
 
     @Override
     public void deleteClient(Long id) {
+
+        if (!clientRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Cliente não encontrado");
+        }
         clientRepository.deleteById(id);
     }
